+++ tests/rpc/index.js
@@ -5,17 +5,17 @@ const client = Neon.create.rpcClient('TestNet')
 describe('RPC methods', function () {
   this.timeout(15000)
 
-  it('test block height', (done) => {
-    client.getBlockCount().then((count) => {
-      count.should.be.a('number')
-      done()
-    })
+  it('test block height', () => {
+    return client.getBlockCount()
+      .then((count) => {
+        count.should.be.a('number')
+      })
   })
 
-  it('test storage', (done) => {
-    client.getStorage(Neon.CONST.CONTRACTS.TEST_RPX, 'totalSupply').then((response) => {
-      response.should.be.a('string')
-      done()
-    })
+  it('test storage', () => {
+    return client.getStorage(Neon.CONST.CONTRACTS.TEST_RPX, '9847e26135152874355e324afd5cc99f002acb33')
+      .then((response) => {
+        response.should.be.a('string')
+      })
   })
 })